+++ firmware/setup_env_slac.sh
@@ -1,2 +1,8 @@
 # Setup the Xilinx software and licensing
 source /afs/slac.stanford.edu/g/reseng/xilinx/vivado_2018.2/Vivado/2018.2/settings64.sh
+
+# Setup VCS simulation environment 
+source /afs/slac.stanford.edu/g/reseng/synopsys/vcs-mx/N-2017.12-1/settings.sh
+source /afs/slac.stanford.edu/g/reseng/synopsys/ns/G-2012.06/settings.sh
+source /afs/slac.stanford.edu/g/reseng/synopsys/CosmosScope/D-2010.03/settings.sh
+source /afs/slac.stanford.edu/g/reseng/zeromq/4.2.0/settings.sh
