--- software/config_20190226_171725.yml
@@ -1,98 +1,0 @@
-ClinkDev:
-  enable: True
-  ForceWrite: False
-  ClinkFeb[0]:
-    enable: True
-    AxiVersion:
-      enable: True
-      ScratchPad: 0x000000
-      FpgaReloadHalt: 0x0
-      FpgaReloadAddress: 0x0
-      UserReset: 0x0
-    AxiMicronN25Q:
-      enable: True
-      Test: 0x0
-    BoardTemp:
-      enable: True
-      AlertMaskWrite: False
-      AlertMask: False
-      StopWrite: False
-      Stop: False
-      RcritMaskWrite: False
-      RcritMask: False
-      LcritMaskWrite: False
-      LcritMask: False
-      FaultQueueWrite: False
-      FaultQueue: False
-      ConvertRateWrite: 0x0
-      ConvertRate: '16'
-      LocalHighSetpointWrite: 0x0
-      LocalHighSetpoint: 70
-      LocalLowSetpointWrite: 0x0
-      LocalLowSetpoint: 0
-      RemoteHighSetpointHighByteWrite: 0x0
-      RemoteHighSetpoint: 70
-      RemoteLowSetpointHighByteWrite: 0x0
-      RemoteLowSetpoint: 0
-      RemoteTcritSetpoint: 85
-      LocalTcritSetpoint: 85
-      TcritHysteresis: 10
-      AlertMode: False
-    BoardPwr:
-      enable: True
-      Control: 0xc
-    Xadc:
-      enable: True
-    ClinkTop:
-      enable: True
-      LinkReset: False
-      ChannelA:
-        enable: True
-        LinkMode: Disable
-        DataMode: None
-        FrameMode: None
-        TapCount: 0x0
-        DataEn: False
-        BaudRate: 0
-        SwControlValue: 0x0
-        SwControlEn: 0x0
-      ChannelB:
-        enable: True
-        LinkMode: Disable
-        DataMode: None
-        FrameMode: None
-        TapCount: 0x0
-        DataEn: False
-        BaudRate: 0
-        SwControlValue: 0x0
-        SwControlEn: 0x0
-    TrigCtrl[0]:
-      enable: True
-      EnableTrig: False
-      InvCC: False
-      TrigMap: ChA
-      TrigPulseWidth: 0xfff
-      TrigMask: 0x1
-    TrigCtrl[1]:
-      enable: True
-      EnableTrig: False
-      InvCC: False
-      TrigMap: ChB
-      TrigPulseWidth: 0xfff
-      TrigMask: 0x1
-  Hardware:
-    enable: True
-    AxiPcieCore:
-      enable: True
-      AxiPciePhy:
-        enable: True
-      AxiVersion:
-        enable: True
-        ScratchPad: 0x000000
-        FpgaReloadHalt: 0x0
-        FpgaReloadAddress: 0x0
-        UserReset: 0x0
-      DmaIbAxisMon:
-        enable: True
-      DmaObAxisMon:
-        enable: True
